Add slide on code coverage
