+++ Python/CodeCoverage/.gitignore
@@ -1,0 +1,3 @@
+.coverage
+default_run_report
+no_iter_run_report
