+++ Python/CodeCoverage/README.md
@@ -1,0 +1,8 @@
+# CodeCoverage
+Illustrates how to use a code coverage tool for Python.
+
+## What is it?
+* `prog.py`: a simle Python program with a few execution paths.
+* `functions.py`: two functions that can be called from `prog.py`.
+* `Makefile`: runs the application in two modes, and produces the
+    summary reports, as well as the annotated source code for each run.
