+++ Python/CodeEvaluation/README.md
@@ -12,3 +12,7 @@ What is it?
     Only variables of type `int`, '`float`, `str`, `bool` and `list` will
     be printed.`
 1. `source.py`: sample code to evaluate.
+1. `fib.py`: definition of Fibonacci function.
+1. `fac.py`: definition of the recursive permutation function.
+1. `fac_iter.py`: definition of the iterative permutation function.
+1. `exec.py`: script that illustrates various uses of `exec`.
