+++ Python/CodeEvaluation/fac_iter.py
@@ -1,0 +1,5 @@
+def fac_iter(n):
+    result = 1
+    for i in range(2, n + 1):
+        result *= i
+    return result
