--- Python/Evals/README.md
@@ -1,8 +1,0 @@
-# Evals
-Using `eval` and `exec` for dynamic code evaluation in Python.
-
-## What is it?
-1. `fib.py`: definition of Fibonacci function.
-1. `fac.py`: definition of the recursive permutation function.
-1. `fac_iter.py`: definition of the iterative permutation function.
-1. `exec.py`: script that illustrates various uses of `exec`.
