--- Python/Evals/fac.py
@@ -1,5 +1,0 @@
-def fac(n):
-    if n < 2:
-        return 1
-    else:
-        return n*fac(n-1)
