--- Python/Evals/fib.py
@@ -1,5 +1,0 @@
-def fib(n):
-    if n == 0 or n == 1:
-        return 1
-    else:
-        return fib(n-1) + fib(n-2)
